[read] Get colr/cmap closure working with intset
This require gating all the closure code on `std`, which is already the
case for GSUB.
